Add files via upload
This is a movie recommendation system created using machine learning. It uses two datasets- credit and movies. It recommends the movies that are similar to the selected movie by attributes like description, cast , crew, genres etc. It basically functions as a movie search bar. This is a principles of artificial intelligence lab project.